+++ saphir_cli/src/docgen/type_info.rs
@@ -1,27 +1,27 @@
 use syn::{Item as SynItem, Type, PathArguments, GenericArgument, Expr, Lit};
-use crate::docgen::{DocGen};
-use std::borrow::Borrow;
 use crate::docgen::crate_syn_browser::File;
+use crate::docgen::utils::find_macro_attribute_flag;
 
+/// Informations about a Rust Type required to create a corresponding OpenApiType
 #[derive(Clone, Debug)]
 pub(crate) struct TypeInfo {
-    pub name: String,
-    pub type_path: Option<String>,
-    pub is_type_serializable: bool,
-    pub is_type_deserializable: bool,
-    pub is_array: bool,
-    pub is_optional: bool,
-    pub min_array_len: Option<u32>,
-    pub max_array_len: Option<u32>,
+    pub(crate) name: String,
+    pub(crate) type_path: Option<String>,
+    pub(crate) is_type_serializable: bool,
+    pub(crate) is_type_deserializable: bool,
+    pub(crate) is_array: bool,
+    pub(crate) is_optional: bool,
+    pub(crate) min_array_len: Option<u32>,
+    pub(crate) max_array_len: Option<u32>,
 }
 
-impl DocGen {
-    pub fn find_type_info<'b>(
-        &self,
+impl TypeInfo {
+    /// Retrieve TypeInfo for a syn::Type found in a crate_syn_browser::File.
+    pub fn new<'b>(
         file: &'b File<'b>,
         t: &Type
     ) -> Option<TypeInfo> {
-        match t.borrow() {
+        match t {
             Type::Path(p) => {
                 if let Some(s) = p.path.segments.first() {
                     let name = s.ident.to_string();
@@ -44,7 +44,7 @@ impl DocGen {
                             }
                         };
 
-                        if let Some(mut type_info) = self.find_type_info(file, t2) {
+                        if let Some(mut type_info) = TypeInfo::new(file, t2) {
                             match name.as_str() {
                                 "Vec" => type_info.is_array = true,
                                 "Option" => type_info.is_optional = true,
@@ -63,10 +63,10 @@ impl DocGen {
                             }
                         });
                         let is_type_serializable = item_attrs.map(|attrs|
-                            self.find_macro_attribute_flag(attrs, "derive", "Serialize")
+                            find_macro_attribute_flag(attrs, "derive", "Serialize")
                         ).unwrap_or_default();
                         let is_type_deserializable = item_attrs.map(|attrs|
-                            self.find_macro_attribute_flag(attrs, "derive", "Deserialize")
+                            find_macro_attribute_flag(attrs, "derive", "Deserialize")
                         ).unwrap_or_default();
                         return Some(TypeInfo { name, type_path, is_type_serializable, is_type_deserializable, is_array: false, is_optional: false, min_array_len: None, max_array_len: None });
                     }
@@ -81,7 +81,7 @@ impl DocGen {
                     _ => None,
                 };
 
-                if let Some(mut type_info) = self.find_type_info(file, a.elem.as_ref()) {
+                if let Some(mut type_info) = TypeInfo::new(file, a.elem.as_ref()) {
                     type_info.is_array = true;
                     type_info.min_array_len = len.clone();
                     type_info.max_array_len = len;
